requirements.txt
Django==2.2.5
django-taggit==1.1.0
model-mommy==1.6.0
Pillow==5.3.0
polls==0.1
python-decouple==3.1
mysqlclient==1.4.4
pytz==2018.7
six==1.11.0